wiki-export - Exporting parts of the NetBeans Wiki
This project downloads and polishes the WikiText data from
http://wiki.netbeans.org, for inclusion in the Apache NetBeans website.

I am including the original raw content in wiki-content/*.xml (mediawiki
xml files) and in wiki-wikimedia (extracted wikitext) to avoid
downloading it everytime the project is run, and for historical
purposes.

The idea is to generate web pages from the data in
wiki-export/wiki-wikimedia files.