[NETBEANS-4378] Extract data for OSGI jars and downloadsize
For OSGI jars, there is no existing info.xml that can be parsed to
determine the data required for the update center catalog. Instead the
code from NetBeans Build System is reused.

The code is written in java and thus an embedded project is created to
hold that code. The java code parses the JAR manifest and localization
bundle and outputs an XML stream, that matches the info.xml structure.

As the artifact is already downloaded, the download size is stored to
provide the value in the catalog.

It is expected, that java binary is on the path, that can execute the
extractor code.