[NETBEANS-4987] Decouple building plugin catalog from UI actions
Building the plugin catalog can potentially take much time and errors
from rebuilding can be caused outside the action area of the executing
user.

This PR implements a different strategy: The plugin catalogs are rebuild
by a CLI command, which is expected to be run periodically by a cron
job. Actions that cause the plugin catalog to change just make an entry
into the netbeans version table and the CLI tools check if catalog
rebuilding is required.

To ensure the catalog is useable, plugins that fail verification are
automatically excluded from the catalog. So instead of failing to
completely rebuild the catalog, a partitial catalog will be created.