[NETBEANS-4987] Show details for errors from catalog generation in admin area
When catalog regeneration is requested from the admin area, the
verification errors for all plugins are displayed to the administrator.
This can help to pin point problems in the building process.

The error message includes the names and ids of the problematic plugins
and the details messages of the verification errors.