Add E-Mail sending option for Pluginportal Admins
Portal Admin gain a GUI to formulate an email text, which can then be
sent to the author of the plugins based on the selection of the modules
and their verification status.

A preview can be created, which also allows to check the size of the
target group as a minimal statistics about the theoretical target group
is created.