PP3: Allow requesting a new validation without first dropping an existing validation
The plugin portal requires a validation of new versions of a plugin.
Before this change a request for a new validation required an existing
validition to be dropped first. For the timeframe between the old
validation being droppend and the revalidation the plugin is not
available to users.

This change modifies the behavior of the portal, so that a revalidation
can be requested while one exists. Old validations are dropped once a
new validation is done.