PP3: Don't require reverification of already verified plugin version
Before this change when compatibilty with a new NetBeans version was
to be added a new verification was required. This change first checks
whether there is already an existing verification present for the
plugin version. If there is an existing verification, this is
transfered to the new NB version.

The idea:

- the author is responsible for the general usability of the plugin
- verifiers check that the plugin basicly does what is promised and
  behaves in a sane way and does no harm.
- NB promises safe updates and thus general plugin behavior should not
  be affected by updates and if so should be caught by the author.