Add 150 more SVG icons (April 2025 batch), and new artboard mappings.
This commit updates the Adobe Illustrator file that holds all of NetBeans' custom-drawn SVG icons, to include ~150 new icons that will be integrated into the main NetBeans repo in a separate PR.

Various new mappings of PNG/GIF file paths to artboard names are also included, and new artboards, with original bitmap icons in a separate layer, were created in the Illustrator file for icons that need to be drawn in the future. New artboards were added in batch using a script that was generated by IconTasks.java.

See the corresponding PR in the main NetBeans repo for screenshots of the new icons.